remove callbacks and put stuff into app
