+++ app.py
@@ -1,7 +1,6 @@
-from matscholar_web.callbacks import app
+from matscholar_web.app import app
 
 server = app.server
 
 if __name__ == '__main__':
-    print("starting...")
     app.run_server(debug=True)
